+++ console/src/features/DataRelationships/DatabaseRelationshipsTab.stories.tsx
@@ -1,36 +1,23 @@
 import React from 'react';
-import { rest } from 'msw';
 import { ReactQueryDecorator } from '@/storybook/decorators/react-query';
 import { ComponentMeta } from '@storybook/react';
 import { NormalizedTable } from '@/dataSources/types';
 
 import { DatabaseRelationshipsTab } from './DatabaseRelationshipsTab';
-import {
-  metadata,
-  relationshipQueryResponse,
-} from '../RelationshipsTable/DatabaseRelationshipsTable/mocks';
-
-const url = 'http://localhost:8080';
+import { handlers } from './__mocks__';
 
 export default {
   title: 'Data Relationships/Database Relationships Tab',
   component: DatabaseRelationshipsTab,
   decorators: [ReactQueryDecorator()],
   parameters: {
-    msw: [
-      rest.post(`${url}/v1/metadata`, (_req, res, ctx) =>
-        res(ctx.json(metadata))
-      ),
-      rest.post(`${url}/v2/query`, (_req, res, ctx) =>
-        res(ctx.json(relationshipQueryResponse))
-      ),
-    ],
+    msw: handlers(),
   },
 } as ComponentMeta<typeof DatabaseRelationshipsTab>;
 
-const tableSchema: NormalizedTable = {
+const table: NormalizedTable = {
   table_schema: 'public',
-  table_name: 'Album',
+  table_name: 'user',
   table_type: 'TABLE',
   is_table_tracked: true,
   columns: [],
@@ -40,5 +27,10 @@ const tableSchema: NormalizedTable = {
 };
 
 export const Primary = () => (
-  <DatabaseRelationshipsTab tableSchema={tableSchema} currentSource="default" />
+  <DatabaseRelationshipsTab
+    table={table}
+    currentSource="default"
+    migrationMode
+    driver="postgres"
+  />
 );