+++ console/src/features/DataRelationships/components/Form/Form.tsx
@@ -1,15 +1,20 @@
 import { Button } from '@/new-components/Button';
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { CardRadioGroup } from '@/new-components/CardRadioGroup';
+import { DbToRsForm } from '@/features/RemoteRelationships';
+import { RowData } from '@/features/RelationshipsTable';
+import { DataTarget } from '@/features/Datasources';
+import { DbToRemoteSchemaRelationship } from '@/features/MetadataAPI';
+import { Driver } from '@/dataSources';
 import { LocalRelationshipWidget } from '../LocalDBRelationshipWidget';
 import { RemoteDBRelationshipWidget } from '../RemoteDBRelationshipWidget';
 
-type Value =
+type RelOption =
   | 'Local Relationship'
   | 'Remote Database Relationship'
   | 'Remote Schema Relationship';
 
-const data: { value: Value; title: string; body: string }[] = [
+const data: { value: RelOption; title: string; body: string }[] = [
   {
     value: 'Local Relationship',
     title: 'Local Relationship',
@@ -27,12 +32,49 @@ const data: { value: Value; title: string; body: string }[] = [
   },
 ];
 
-export const Form = () => {
-  const [option, setOption] = useState('Local Relationship');
+interface FormProps {
+  existingRelationship?: RowData;
+  sourceTableInfo: DataTarget;
+  driver: Driver;
+  onComplete: (v: {
+    title?: string;
+    message?: string;
+    type: 'success' | 'error' | 'cancel';
+  }) => void;
+}
+
+const getExistingRelationshipType = (
+  existingRelationship?: RowData
+): RelOption => {
+  if (!existingRelationship) return 'Local Relationship';
+  if (existingRelationship.toType === 'database')
+    return 'Remote Database Relationship';
+  if (existingRelationship.toType === 'remote_schema')
+    return 'Remote Schema Relationship';
+
+  return 'Local Relationship';
+};
+
+export const Form = ({
+  existingRelationship,
+  sourceTableInfo,
+  onComplete,
+  driver,
+}: FormProps) => {
+  const [option, setOption] = useState<RelOption>(
+    getExistingRelationshipType(existingRelationship)
+  );
+
+  useEffect(() => {
+    setOption(getExistingRelationshipType(existingRelationship));
+  }, [existingRelationship]);
+
   return (
-    <div className="w-full sm:w-9/12 bg-white shadow-sm rounded p-md border border-gray-300 shadow show">
+    <div className="w-full sm:w-9/12 bg-white shadow-sm rounded p-md border border-gray-300 show">
       <div className="flex items-center mb-md">
-        <Button size="sm">Cancel</Button>
+        <Button size="sm" onClick={() => onComplete({ type: 'cancel' })}>
+          Cancel
+        </Button>
         <span className="font-semibold text-muted ml-1.5">
           Create New Relationship
         </span>
@@ -46,25 +88,41 @@ export const Form = () => {
           items={data}
           onChange={relType => setOption(relType)}
           value={option}
+          disabled={!!existingRelationship}
         />
         {option === 'Local Relationship' ? (
           <LocalRelationshipWidget
-            sourceTableInfo={{
-              database: 'default',
-              schema: 'public',
-              table: 'resident',
-            }}
+            sourceTableInfo={sourceTableInfo}
+            existingRelationshipName={existingRelationship?.name}
+            onComplete={onComplete}
+            driver={driver}
           />
         ) : option === 'Remote Database Relationship' ? (
           <RemoteDBRelationshipWidget
-            sourceTableInfo={{
-              database: 'default',
-              schema: 'public',
-              table: 'resident',
-            }}
+            onComplete={onComplete}
+            sourceTableInfo={sourceTableInfo}
+            existingRelationshipName={existingRelationship?.name}
           />
         ) : (
-          <>do something for remote schema relationships</>
+          <DbToRsForm
+            sourceTableInfo={sourceTableInfo}
+            onComplete={onComplete}
+            selectedRelationship={
+              existingRelationship && {
+                name: existingRelationship.name,
+                definition: {
+                  to_remote_schema: {
+                    lhs_fields: (existingRelationship.relationship as DbToRemoteSchemaRelationship) // asserting type becase here RS relationship is the only possibility, DB relationships are filtered out above with the help of `option`
+                      .lhs_fields, // existing relationship would handle the legacy/ new shape, ie. lhs_fields will always be present (legacy hasura_field is already translated to the new format)
+                    remote_field: (existingRelationship.relationship as DbToRemoteSchemaRelationship)
+                      .remote_field,
+                    remote_schema: (existingRelationship.relationship as DbToRemoteSchemaRelationship)
+                      .remoteSchemaName,
+                  },
+                },
+              }
+            }
+          />
         )}
       </div>
     </div>