+++ console/src/features/MetadataAPI/hooks/useObjectRelationships.ts
@@ -7,6 +7,7 @@ export const useObjectRelationships = (target: DataTarget) => {
     MetadataSelector.getLocalDBObjectRelationships(target.database, {
       name: target.table,
       schema: 'schema' in target ? target.schema : '',
+      ...('dataset' in target && { dataset: target.dataset }),
     })
   );
   return { data, ...rest };