+++ console/src/features/RemoteRelationships/index.ts
@@ -1,3 +1,5 @@
 export * from './RemoteSchemaRelationships/components/RemoteSchemaToRemoteSchemaForm/RemoteSchemaToRemoteSchemaForm';
 export * from './RemoteSchemaRelationships/components/RemoteSchemaToRemoteSchemaForm/RelationshipTypeCardRadioGroup';
 export * from './RemoteSchemaRelationships/components/RemoteSchemaToDB';
+export * from './RemoteDBRelationships/DbToRSRelationships/DbToRsForm';
+export * from './RemoteSchemaRelationships/types';